backend/requirements.txt
# InfoGenie 后端依赖包
# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# 数据库
Flask-PyMongo==2.3.0
pymongo==4.5.0

# 密码加密
Werkzeug==2.3.7

# JWT认证
PyJWT==2.8.0

# HTTP请求
requests==2.31.0

# 邮件发送
Flask-Mail==0.9.1

# 数据处理
python-dateutil==2.8.2

# 环境变量
python-dotenv==1.0.0

# 开发工具
flask-limiter==3.5.0  # API限流
